+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # Next release
 - Show usernames in normal letter case
 - Show event author when viewing specific event
+- Show message if a post or comment has been edited